+++ packages/demoreact/src/app/ChatSupportTest.tsx
@@ -1,7 +1,8 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 import { Web3Provider } from '@ethersproject/providers';
-import { Chat } from '@pushprotocol/uiweb';
-import { Web3Context } from './context';
+import { Chat, ITheme } from '@pushprotocol/uiweb';
+import { lightTheme } from '@pushprotocol/uiweb';
+import { EnvContext, Web3Context } from './context';
 
 export type ChatProps = {
   provider: Web3Provider;
@@ -10,17 +11,32 @@ export type ChatProps = {
   modalTitle?: string;
   primaryColor?: string;
   apiKey?: string;
-  env?: string;
+  env?: "dev" | "staging" | "prod";
 };
 
 export const ChatSupportTest = () => {
   const { account } = useContext<any>(Web3Context);
+  const { env } = useContext<any>(EnvContext);
+  const theme: ITheme = {
+    bgColorPrimary: 'gray',
+    bgColorSecondary: 'purple',
+    textColorPrimary: 'white',
+    textColorSecondary: 'green',
+    btnColorPrimary: 'red',
+    btnColorSecondary: 'purple',
+    border: '1px solid black',
+    borderRadius: '40px',
+    moduleColor: 'pink',
+  };
+
   return (
     <Chat
-    account={account}
-      supportAddress="0xd9c1CCAcD4B8a745e191b62BA3fcaD87229CB26d"
+      account={account}
+      supportAddress="0xD2A371cAFBE09333F6668229E4BD33196583c325"
       apiKey="tAWEnggQ9Z.UaDBNjrvlJZx3giBTIQDcT8bKQo1O1518uF1Tea7rPwfzXv2ouV5rX9ViwgJUrXm"
-      env='dev'
+      env={env}
+      greetingMsg="How can i help you?"
+      theme={lightTheme}
     />
   );
 };