+++ packages/restapi/src/lib/channels/index.ts
@@ -4,3 +4,4 @@ export * from './search';
 export * from './subscribe';
 export * from './unsubscribe';
 export * from './_getSubscribers';
+export * from './getSubscribers';