+++ packages/restapi/src/lib/channels/unsubscribe.ts
@@ -9,13 +9,13 @@ import {
   getDomainInformation,
   getSubscriptionMessage
 } from './signature.helpers';
-import Constants from '../constants';
+import Constants, {ENV} from '../constants';
 
 type SignerType = {
   _signTypedData: (
-    domain: unknown,
-    types: unknown,
-    value: unknown
+    domain: any,
+    types: any,
+    value: any
   ) => Promise<string>
 }
  
@@ -24,7 +24,7 @@ export type UnSubscribeOptionsType = {
   channelAddress: string;
   userAddress: string;
   verifyingContractAddress?: string;
-  env?: string;
+  env?: ENV;
   onSuccess?: () => void
   onError?: (err: Error) => void,
 }