+++ packages/restapi/src/lib/chat/approveRequest.ts
@@ -1,48 +1,75 @@
 import axios from 'axios';
-import {
-  getAPIBaseUrls,
-} from '../helpers';
+import { getAPIBaseUrls } from '../helpers';
 import Constants from '../constants';
-import { AccountEnvOptionsType } from '../types';
-import { approveRequestPayload, createUserIfNecessary, IApproveRequestPayload } from './helpers';
+import { EnvOptionsType, SignerType } from '../types';
+import { approveRequestPayload, sign, getConnectedUser, IApproveRequestPayload, getAccountAddress, getWallet, createUserIfNecessary } from './helpers';
+import CryptoES from "crypto-es"
 
-/**
- *  POST '/v1/chat/request/accept
- */
+interface ApproveRequestOptionsType extends EnvOptionsType {
+  /**
+   * Chat request sender address
+   */
+  senderAddress: string;
+  // pgpPrivateKey?: string;
 
-interface ApproveRequestOptionsType extends AccountEnvOptionsType {
-  senderAddress: string; // chat request sender address
-  // privateKey?: string; // private key for signature
+  /**
+   * Request state. As of now, only `Approved` is allowed
+   */
   status?: 'Approved';
   // sigType?: string;
+  account?: string;
+  signer?: SignerType;
 }
 
+/**
+ * Approve Chat Request
+ */
 export const approve = async (
   options: ApproveRequestOptionsType
 ) => {
   const {
     status = "Approved",
     // sigType = 'sigType',
-    // privateKey = null,
-    account,
+    account = null,
+    signer = null,
     senderAddress,
     env = Constants.ENV.PROD,
+    // pgpPrivateKey = null,
   } = options || {};
 
-  // get user with raw privateKey
-  // const createdUser: IConnectedUser = await getConnectedUser(account, privateKey, env);
-  const connectedUser = await createUserIfNecessary({account,env});
-  // TODO: make signature
+  if(account == null && signer == null) {
+    throw new Error(`At least one from account or signer is necessary!`);
+  }
+
+  const wallet = getWallet({ account, signer });
+  const address = await getAccountAddress(wallet);
+
   const API_BASE_URL = getAPIBaseUrls(env);
   const apiEndpoint = `${API_BASE_URL}/v1/chat/request/accept`;
 
-  const requestUrl = `${apiEndpoint}`;
+  
 
-  const body:IApproveRequestPayload = approveRequestPayload(senderAddress,account,status);
+  const body: IApproveRequestPayload = approveRequestPayload(senderAddress, address, status);
 
-  return axios.put(requestUrl, body)
+  // const bodyToBeHashed = {
+  //   fromDID: body.fromDID,
+  //   toDID: body.toDID,
+  //   status: body.status
+  // }
+
+  const connectedUser = await createUserIfNecessary(wallet,env);
+  // const hash = CryptoES.SHA256(JSON.stringify(bodyToBeHashed)).toString()
+  // const signature: string = await sign({
+  //   message: hash,
+  //   signingKey: connectedUser.privateKey!
+  // });
+  // const sigType = "pgp";
+  // const verificationProof: string = sigType + ":" + signature;
+  // body.verificationProof = verificationProof;
+
+  return axios.put(apiEndpoint, body)
     .catch((err) => {
-      console.error(`[EPNS-SDK] - API ${requestUrl}: `, err);
-      throw Error(`[EPNS-SDK] - API ${requestUrl}: ${err}`);
+      console.error(`[Push SDK] - API ${approve.name}: `, err);
+      throw Error(`[Push SDK] - API ${approve.name}: ${err}`);
     });
 }