+++ packages/restapi/src/lib/helpers/address.ts
@@ -1,5 +1,5 @@
 import * as ethers from 'ethers';
-import Constants from '../constants';
+import Constants, { ENV } from '../constants';
 
 export interface AddressValidatorsType {
   [key: string]: ({ address } : { address: string }) => boolean;
@@ -10,6 +10,12 @@ export function isValidETHAddress(address: string) {
     return ethers.utils.isAddress((address.split(':'))[1]);
   }
   return ethers.utils.isAddress(address);
+}
+
+export function  isValidNFTCAIP10Address (realCAIP10: string)  {
+        const walletComponent = realCAIP10.split(':');
+        if (isNaN(Number(walletComponent[1]))) return false
+        return (walletComponent.length === 3 && walletComponent[0] === 'eip155' && ethers.utils.isAddress(walletComponent[2]))
 }
 
 const AddressValidators: AddressValidatorsType = {
@@ -60,7 +66,7 @@ export function getCAIPDetails(addressInCAIP: string) : CAIPDetailsType | null {
   return null;
 }
 
-export function getFallbackETHCAIPAddress(env: string, address: string) {
+export function getFallbackETHCAIPAddress(env: ENV, address: string) {
   let chainId = 1; // by default PROD
 
   if (env === Constants.ENV.DEV || env === Constants.ENV.STAGING) {
@@ -80,7 +86,7 @@ export function getFallbackETHCAIPAddress(env: string, address: string) {
  *    else 
  *      throw error!
  */
-export function getCAIPAddress(env: string, address: string, msg?: string) {
+export function getCAIPAddress(env: ENV, address: string, msg?: string) {
   if (validateCAIP(address)) {
     return address;
   } else {