+++ packages/uiweb/src/lib/components/chat/index.tsx
@@ -1,21 +1,23 @@
 import React, { useEffect, useState } from 'react';
 import ChatIcon from '../../icons/chat/chatIcon.svg';
 import { Modal } from './Modal';
-import Constants from './constants';
 import styled from 'styled-components';
 import { handleOnChatIconClick } from '../../helpers';
 import { ChatMainStateContext, ChatPropsContext } from '../../context';
-import { IMessageIPFS } from '@pushprotocol/restapi';
+import { IMessageIPFS, ITheme } from '../../types';
 import './index.css';
+import { Constants, ENV, lightTheme } from '../../config';
+import { useSDKSocket } from '../../hooks/useSDKSocket';
+
 
 export type ChatProps = {
   account: string;
   supportAddress: string;
-  // greetingMsg?: string;
+  greetingMsg?: string;
   modalTitle?: string;
-  primaryColor?: string;
+  theme?: ITheme;
   apiKey?: string;
-  env?: string;
+  env?: ENV;
 };
 
 export type ButtonStyleProps = {
@@ -25,9 +27,9 @@ export type ButtonStyleProps = {
 export const Chat: React.FC<ChatProps> = ({
   account,
   supportAddress,
-  // greetingMsg = Constants.DEFAULT_GREETING_MSG,
+  greetingMsg = Constants.DEFAULT_GREETING_MSG,
   modalTitle = Constants.DEFAULT_TITLE,
-  primaryColor = Constants.COLOR.PRIMARY,
+  theme = { ...lightTheme },
   apiKey = '',
   env = Constants.ENV.PROD,
 }) => {
@@ -35,22 +37,30 @@ export const Chat: React.FC<ChatProps> = ({
   const [connectedUser, setConnectedUser] = useState<any>(null);
   const [messageBeingSent, setMessageBeingSent] = useState<boolean>(false);
   const [message, setMessage] = useState<string>('');
-  const [footerError, setFooterError] = useState<string>('');
+  const [toastMessage, setToastMessage] = useState<string>('');
+  const [toastType, setToastType] = useState<'error' | 'success'>();
   const [chats, setChats] = useState<IMessageIPFS[]>([]);
 
   const setChatsSorted = (chats: IMessageIPFS[]) => {
-    chats.sort((a, b) => {
+    const uniqueChats = [...new Map(chats.map((item) => [item["timestamp"], item])).values()];
+    
+    uniqueChats.sort((a, b) => {
       return a.timestamp! > b.timestamp! ? 1 : -1;
     });
-    setChats(chats);
+    setChats(uniqueChats);
   };
+  const socketData = useSDKSocket({
+    account: account,
+    env,
+    apiKey,
+  });
 
   const chatPropsData = {
     account,
     supportAddress,
-    // greetingMsg,
+    greetingMsg,
     modalTitle,
-    primaryColor,
+    theme: { ...lightTheme, ...theme },
     apiKey,
     env,
   };
@@ -60,20 +70,27 @@ export const Chat: React.FC<ChatProps> = ({
     setConnectedUser(null);
   }, [account, supportAddress]);
 
+
+  
   const chatMainStateData = {
     isModalOpen,
+    socketData,
     setIsModalOpen,
     connectedUser,
     setConnectedUser,
     messageBeingSent,
     setMessageBeingSent,
+    setToastMessage,
+    setToastType,
     message,
     setMessage,
     chats,
     setChatsSorted,
-    footerError,
-    setFooterError,
+    toastMessage,
+    toastType,
   };
+
+  
 
   return (
     <Container>
@@ -81,7 +98,7 @@ export const Chat: React.FC<ChatProps> = ({
         <ChatMainStateContext.Provider value={chatMainStateData}>
           {!isModalOpen && (
             <Button
-              bgColor={primaryColor}
+              bgColor={theme.btnColorPrimary!}
               onClick={() =>
                 handleOnChatIconClick({ isModalOpen, setIsModalOpen })
               }