+++ packages/uiweb/src/lib/config/themisation.tsx
@@ -1,0 +1,15 @@
+
+import { ITheme } from "../types"
+
+
+export const lightTheme: ITheme = {
+  bgColorPrimary: "#fff",
+  bgColorSecondary: "#D53A94",
+  textColorPrimary: "#1e1e1e",
+  textColorSecondary: "#fff",
+  btnColorPrimary: "#D53A94",
+  btnColorSecondary: "#494D5F",
+  border: "1px solid #E4E8EF",
+  borderRadius:"24px",
+  moduleColor:"#fff",
+}