+++ packages/uiweb/CHANGELOG.md
@@ -2,6 +2,57 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.3](https://github.com/ethereum-push-notification-service/push-sdk/compare/uiweb-0.4.2...uiweb-0.4.3) (2023-02-23)
+
+
+### Bug Fixes
+
+* fixed image width in notifs ([#196](https://github.com/ethereum-push-notification-service/push-sdk/issues/196)) ([a4100d2](https://github.com/ethereum-push-notification-service/push-sdk/commit/a4100d26e0036f717d7a1846285901d8e3badb81))
+
+
+
+## [0.4.2](https://github.com/ethereum-push-notification-service/push-sdk/compare/uiweb-0.4.1...uiweb-0.4.2) (2023-02-17)
+
+
+### Bug Fixes
+
+* fixed message returned from socket ([#168](https://github.com/ethereum-push-notification-service/push-sdk/issues/168)) ([ee6fd04](https://github.com/ethereum-push-notification-service/push-sdk/commit/ee6fd04d781c2639885a8eef6b996ca945829028))
+
+
+
+## [0.4.1](https://github.com/ethereum-push-notification-service/push-sdk/compare/uiweb-0.4.0...uiweb-0.4.1) (2023-02-03)
+
+
+### Bug Fixes
+
+* converted svg to react component ([#135](https://github.com/ethereum-push-notification-service/push-sdk/issues/135)) ([e68c783](https://github.com/ethereum-push-notification-service/push-sdk/commit/e68c78386adfebecad8bcba514a9272f008bdc4f))
+* Metamask reject handled ([#136](https://github.com/ethereum-push-notification-service/push-sdk/issues/136)) ([46a20b7](https://github.com/ethereum-push-notification-service/push-sdk/commit/46a20b751e655a526168c8ae7c400211f3b62b15))
+
+
+
+# [0.4.0](https://github.com/ethereum-push-notification-service/push-sdk/compare/uiweb-0.3.0...uiweb-0.4.0) (2023-01-27)
+
+
+### Bug Fixes
+
+* fixed graph svg in notifs ([#129](https://github.com/ethereum-push-notification-service/push-sdk/issues/129)) ([a653300](https://github.com/ethereum-push-notification-service/push-sdk/commit/a653300202dfeb85951d368f968e1859fefea7b5))
+
+
+### Features
+
+* merge chat-extra-features ([74268c6](https://github.com/ethereum-push-notification-service/push-sdk/commit/74268c6712615d08a2fd08325a11e4ab805905ac))
+
+
+
+# [0.3.0](https://github.com/ethereum-push-notification-service/push-sdk/compare/uiweb-0.2.3...uiweb-0.3.0) (2023-01-13)
+
+
+### Features
+
+* Bsc chain config changes added ([#113](https://github.com/ethereum-push-notification-service/push-sdk/issues/113)) ([e1af11b](https://github.com/ethereum-push-notification-service/push-sdk/commit/e1af11b1fa444e30f8aa08ee6b54a30bb03d6070))
+
+
+
 ## [0.2.3](https://github.com/ethereum-push-notification-service/push-sdk/compare/uiweb-0.2.2...uiweb-0.2.3) (2022-11-04)